data/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
arrow==0.17.0
    # via -r requirements.in
attrs==20.3.0
    # via pytest
azure-common==1.1.26
    # via azure-mgmt-resource
azure-core==1.11.0
    # via
    #   azure-cosmos
    #   azure-eventhub
    #   azure-mgmt-core
    #   azure-storage-blob
    #   azure-storage-file-datalake
azure-cosmos==4.2.0
    # via -r requirements.in
azure-eventhub==5.3.0
    # via -r requirements.in
azure-mgmt-core==1.2.2
    # via azure-mgmt-resource
azure-mgmt-resource==15.0.0
    # via -r requirements.in
azure-storage-blob==12.7.1
    # via
    #   -r requirements.in
    #   azure-storage-file-datalake
azure-storage-file-datalake==12.2.3
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   msrest
    #   requests
    #   uamqp
cffi==1.14.5
    # via cryptography
chardet==4.0.0
    # via requests
coverage==5.4
    # via pytest-cov
cryptography==3.4.6
    # via azure-storage-blob
cycler==0.10.0
    # via matplotlib
docopt==0.6.2
    # via -r requirements.in
et-xmlfile==1.0.1
    # via openpyxl
flake8==3.8.4
    # via -r requirements.in
humanize==3.2.0
    # via -r requirements.in
idna==2.10
    # via requests
iniconfig==1.1.1
    # via pytest
isodate==0.6.0
    # via msrest
jdcal==1.4.1
    # via openpyxl
jinja2==2.11.3
    # via -r requirements.in
kiwisolver==1.3.1
    # via matplotlib
markupsafe==1.1.1
    # via jinja2
matplotlib==3.3.4
    # via -r requirements.in
mccabe==0.6.1
    # via flake8
msrest==0.6.21
    # via
    #   azure-mgmt-resource
    #   azure-storage-blob
    #   azure-storage-file-datalake
numpy==1.20.1
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
oauthlib==3.1.0
    # via requests-oauthlib
openpyxl==3.0.6
    # via -r requirements.in
packaging==20.9
    # via pytest
pandas==1.2.2
    # via -r requirements.in
pillow==8.1.0
    # via matplotlib
pipdeptree==2.2.1
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
psutil==5.8.0
    # via -r requirements.in
py==1.10.0
    # via pytest
pycodestyle==2.6.0
    # via flake8
pycparser==2.20
    # via cffi
pyflakes==2.2.0
    # via flake8
pymongo==3.11.3
    # via -r requirements.in
pyparsing==2.4.7
    # via
    #   matplotlib
    #   packaging
pytest==6.2.2
    # via
    #   -r requirements.in
    #   pytest-cov
pytest-cov==2.11.1
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   arrow
    #   matplotlib
    #   pandas
pytz==2021.1
    # via
    #   -r requirements.in
    #   pandas
redis==3.5.3
    # via -r requirements.in
requests==2.25.1
    # via
    #   -r requirements.in
    #   azure-core
    #   msrest
    #   requests-oauthlib
requests-oauthlib==1.3.0
    # via msrest
six==1.15.0
    # via
    #   azure-core
    #   azure-cosmos
    #   cycler
    #   isodate
    #   python-dateutil
    #   uamqp
toml==0.10.2
    # via pytest
uamqp==1.2.14
    # via azure-eventhub
urllib3==1.26.3
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
